abstract out format utils
